hpscreg: fix missing return in parse_supplements()
When running "import hpscreg" on a fresh instance
(where only a db_import was done)
we could reach a state where parse_supplements
doesn't return a value, equivalent with returning None
and leading to:

TypeError: 'NoneType' object is not iterable

This is now fixed.
We always want to return some type of list,
possible the empty dirty_supplements.