proper status for free_text only diseases
On hPSCreg it is allowed to have only a free_text
when there is no suitable purl for a disease.

In this case we use the free_text instead
of displaying "Disease: no".
We don't show the free_text again as "notes",
which is done for diseases with a purl.