upgrade requests for recent CAs
starting with python-requests 2.16
there is a separate certify library
that can be upgraded to get newer certificates.

We were missing trusted certificates for www.ebi.ac.uk,
so we upgrade to the latest patch build of requests 2.16
and the latest certify.

This solves
SSLError: SSL: CERTIFICATE_VERIFY_FAILED
for www.ebi.ac.uk
and should provide an easy route to update
the python certificate store via certifi.