add at_ecacc in export API
for hPSCreg to use

available_for_sale is the "go live" flag,
set to yet when they are inteded to be sold.
For withdrawn lines this is False,
but in the API it is explicitly
set to True again for ECACC use.

hPSCreg needs to know that the line
is not available anymore,
so we export available_for_sale_at_ecacc as at_ecacc

We also display now also the go live flag in the line status
in the detail line pages in the executive dashboard.
This might help for debugging later.

refs #8726